medvision/backend/requirements.txt
google-adk>=0.5.0
google-genai>=1.5.0
google-cloud-firestore>=2.13.0
google-cloud-storage>=2.10.0
google-cloud-logging>=3.8.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
